Refactor color handling in custom painters for consistency and precision
- Updated color definitions in various custom painter classes to use double values for RGB and alpha components, enhancing color accuracy.
- Modified PlayerShipPainter, EnemyShipPainter, and other related classes to ensure consistent color representation across the application.
- Improved maintainability by standardizing color handling methods, contributing to a more cohesive visual design.

+++ lib/widgets/background.dart
@@ -106,7 +106,10 @@ class StarPainter extends CustomPainter {
 
     for (var star in stars) {
       paint.color = Colors.white.withValues(
-          red: 255, green: 255, blue: 255, alpha: star.opacity * 255);
+          red: 255.0,
+          green: 255.0,
+          blue: 255.0,
+          alpha: (star.opacity * 255).toDouble());
       canvas.drawCircle(star.position, star.size, paint);
     }
   }
